Improved installation proccess and adding sports marker set #34632
