+++ Templates/run_opensim_example.bat
@@ -1,0 +1,7 @@
+
+call conda activate opensim_example
+
+python "%~dp0qtm2opensim.py" --c3d_dir "%~1"
+
+
+exit
