+++ requirements.txt
@@ -1,0 +1,40 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: win-64
+blas=1.0=mkl
+ca-certificates=2024.3.11=haa95532_0
+contourpy=1.1.1=pypi_0
+cycler=0.12.1=pypi_0
+fonttools=4.51.0=pypi_0
+freeglut=3.0.0=h6538335_5
+importlib-resources=6.4.0=pypi_0
+intel-openmp=2023.1.0=h59b6b97_46320
+kiwisolver=1.4.5=pypi_0
+libffi=3.4.4=hd77b12b_0
+matplotlib=3.7.5=pypi_0
+mkl=2023.1.0=h6b88ed4_46358
+mkl-service=2.4.0=py38h2bbff1b_1
+mkl_fft=1.3.8=py38h2bbff1b_0
+mkl_random=1.2.4=py38h59b6b97_0
+numpy=1.20.3=py38h749eb61_1
+numpy-base=1.20.3=py38h5bfbeaa_1
+opensim=4.4=py38np120
+openssl=3.0.13=h2bbff1b_0
+packaging=24.0=pypi_0
+pandas=2.0.3=pypi_0
+pillow=10.3.0=pypi_0
+pip=23.3.1=py38haa95532_0
+pyparsing=3.1.2=pypi_0
+python=3.8.19=h1aa4202_0
+python-dateutil=2.9.0.post0=pypi_0
+pytz=2024.1=pypi_0
+scipy=1.10.1=pypi_0
+setuptools=68.2.2=py38haa95532_0
+six=1.16.0=pypi_0
+sqlite=3.41.2=h2bbff1b_0
+tbb=2021.8.0=h59b6b97_0
+tzdata=2024.1=pypi_0
+vc=14.2=h21ff451_1
+vs2015_runtime=14.27.29016=h5e58377_2
+wheel=0.41.2=py38haa95532_0
+zipp=3.18.1=pypi_0
